move ExecCommand under arduino/utils
