+++ legacy/builder/phases/sizer.go
@@ -21,9 +21,9 @@ import (
 	"regexp"
 	"strconv"
 
+	"github.com/arduino/arduino-cli/arduino/builder/utils"
 	"github.com/arduino/arduino-cli/legacy/builder/builder_utils"
 	"github.com/arduino/arduino-cli/legacy/builder/types"
-	"github.com/arduino/arduino-cli/legacy/builder/utils"
 	"github.com/arduino/go-properties-orderedmap"
 	"github.com/pkg/errors"
 )
