--- legacy/builder/test/utils_test.go
@@ -1,42 +1,0 @@
-// This file is part of arduino-cli.
-//
-// Copyright 2020 ARDUINO SA (http://www.arduino.cc/)
-//
-// This software is released under the GNU General Public License version 3,
-// which covers the main part of arduino-cli.
-// The terms of this license can be found at:
-// https://www.gnu.org/licenses/gpl-3.0.en.html
-//
-// You can be released from the requirements of the above licenses by purchasing
-// a commercial license. Buying such a license is mandatory if you want to
-// modify or otherwise use the software for commercial activities involving the
-// Arduino software without disclosing the source code of your own applications.
-// To purchase a commercial license, send an email to license@arduino.cc.
-
-package test
-
-import (
-	"testing"
-
-	"github.com/arduino/arduino-cli/legacy/builder/utils"
-	"github.com/stretchr/testify/require"
-)
-
-func TestPrintableCommand(t *testing.T) {
-	parts := []string{
-		"/path/to/dir with spaces/cmd",
-		"arg1",
-		"arg-\"with\"-quotes",
-		"specialchar-`~!@#$%^&*()-_=+[{]}\\|;:'\",<.>/?-argument",
-		"arg   with spaces",
-		"arg\twith\t\ttabs",
-		"lastarg",
-	}
-	correct := "\"/path/to/dir with spaces/cmd\"" +
-		" arg1 \"arg-\\\"with\\\"-quotes\"" +
-		" \"specialchar-`~!@#$%^&*()-_=+[{]}\\\\|;:'\\\",<.>/?-argument\"" +
-		" \"arg   with spaces\" \"arg\twith\t\ttabs\"" +
-		" lastarg"
-	result := utils.PrintableCommand(parts)
-	require.Equal(t, correct, result)
-}
